del asinkloader from reduser
